Immediately deliver new intent to paused activities.
ag/1437028 tried to fix this problem, but limited intent
delivery to only paused activities in a minimized docked stack
to reduce the effect of the change since we are late in the
release cycle. However, it didn't completely solve the problem
for example when the ChooserActivity comes up in the fullscreen
stack un-minimizing the docked stack and then the user selects
the app in the docked stack from ChooserActivity.
We now immediately deliver intents to paused activities as they
are currently visible to the user and we want the user to see
the visual effects caused by the intent delivery now.

Bug: 31371093
Change-Id: I94f9717da66daa512071bb96e62b2f9811691a78