Save TaskRecord before calling performClearTaskLocked.
If the activity exists in a task that is in persistent storage, the
task reference will be cleared when its brought to the foreground in
this method. Alternatively, if the activity is finished immediately
as a result of clearing task, the task reference will be lost.

Therefore, we must persist the task reference before calling it.

Change-Id: Id38cf6654f5d4f5b727441048dc6b9a7568b574e
Fixes: 35868563
Test: manual (code and bugreport inspection)
(cherry picked from commit 0966c9e59ac2df9ddc49c67bf69126377b2046b4)