Work on issue #36891897: Need to ensure foreground services...
...can't hide themselves

Propagate to notification manager the apps that are causing
the "running in background" notification to be shown.

Also hopefully this time fix the problem with the notification
being stuck.  (We were mixing elapsed time in the state management
with uptime in the message scheduling.)

Test: manual

Change-Id: I9e38bff5fe69fa75b418e34c84d4e704ef70f460
(cherry picked from commit cd0f448ced81d034fe895b6d25ad4828c7cfd7db)