Fixes the animation and state when docking a task by drag
There was 3 issues in this bug.

First the wrong stack id was used to defer and continue updating bounds
that broke a lot of the animation. This was fixed by using the recents
stack id instead of home.

Secondly, the wrong transit was passed into AppTransition.java getting
the wrong animation for recents to be docked, defering surface layouts
ensures that the correct transition is gained in that transaction.

Lastly, remove the starting window for the docked-by-drag-from-recents
transition because it was causing the window container to freeze its
bounds causing clipping issues and 2 flickers (where it appears over
the thumbnail) during the animation. Added a TODO to fix it later.

Test: go/wm-smoke
Bug: 34099271
Change-Id: Iaf0dffe5c2f5108c9946c9ea23c6e3fd6a49c34d