Fix flickery behavior when launching notification when occluded
- Launch the action after the panel is fully collapses in this
case such that we have a clean transition, like we do when
launching all other activities from the shade.
- When Keyguard flags changed, don't execute the app transition
when there was already be one before, as it might interfere with
that one and start it prematurely.

Test: go/wm-smoke
Test: Launch notification from occluded notification shade
Bug: 63686192
Change-Id: I49efb2f627fcc7a0d6386dcea2510d31ab520b11
(cherry picked from commit e7860e392de82567365b385b522eae48c36f21c0)