Move home stack behind top fullscreen stack when split-screen is dismissed
Further improvement on ag/3245389 where we were moving the home stack
behind the top stack, however we don't want to move the home stack
behind a top stack that isn't fullscreen like Pip.
Also:
- Fixed issue where we are creating a non-resizeable stack in
split-screen first because we were resolving the windowing mode again in
setWindowingMode which doesn't take into account the activity we will be
placing into the stack.
- Don't ensure visible activities if a stack windowing mode is changing
as a side effect of us entering split-screen mode.
- Make sure recents activity is moved to the front if it already
exists when we are entering split-screen mode.

Change-Id: I07a681c74713abb3eef01c298ee15c5752212cb2
Fixes: 69980806
Fixes: 69662547
Fixes: 69898573
Test: go/wm-smoke
Test: Steps from bug.