Restore bounds w/o adjusting it.
Looking at existing code in updateOverrideConfiguration(Rect), it
should be fine to skip them at restoring stage because:
1) If bounds is null/empty, it unsets Configuration. This task was only
restored from XML so the Configuration is for sure empty;
2) If bounds is concrete, it adjusts bounds to allowed values based on
context -- minHeight/minWidth and ActivityStack. If the logic is correct
there isn't need to adjust it based on minimal dimensions because it was
restored, while there is not an ActivityStack at that time, so we have
no choice but to delay that adjustment to when it is added to an
ActivityStack.

Note this is a rather short term solution, because there is not a
complete story about persisting Configuration yet, which may be
necessary for a better support of freeform tasks -- such as previous
windowing mode and display ID etc.

Bug: 70573444
Test: go/wm-smoke w/o PiP on sailfish. Task restoration works on ARC
branch.

Change-Id: Ie2cfce6bb8ef7b521316253268d788ecf746eabf